Cut the gordian knot by figuring all Features can have some text to retrieve from them, fleshed out ConfigWindow and added logic piping from ConfigMenu into ConfigPage/Window
All Features have a String return, so user submitted data or data that is displayed (potentially that is pulled from a config property) can be grabbed via Feature identifier, added draw() logic for most Features (still need FeatureComposite figured out), ConfigWindow exists now as the keeper of the WindowFrame, manager of its ConfigPages, and bearer of a sliding tab of selectable active Pages (ConfigSelectorPanel). Most of the pipelining is present to instantiate this and test it, just need the draw command for FeatureComposite to be able to see a basic version of it in action. New UML diagram, too.