Added text input, cleaned up and fixed proper the property updating, considerable auto-formatting to ease usage
FeatureTextInput is now here for custom text input, the property updating behavior works, ConfigMenu has a method for establishing the properties in a particular config file and also has more integration with the ConfigBlueprint stuff from SVI, added handling of user adding Features with multiple vertical proportionality or in the middle of buffered space to account for what is most likely intended and to ease the overall process of using this (if you want something at column 3, it should be at column 3, not at column 0 if the row was empty; if you want something at column 2 but there was spacing there, it just subsumes the spacing and puts itself there). Having some teething issues with the SVI overhaul I'm working on that's causing features that were working to break, so needs some work still.